refactor(teaminbox): extract TeamInboxView pagination and detail
Move page loading/refresh into useTeamInboxPagination, read transitions
into useTeamInboxReadActions, the right pane into
components/TeamInboxDetailPane, and loadStateForPage plus the work-item
row reconciler into domain/. Behaviour unchanged.
